Re-factored the NUKE code into a gait engine class and added a wheel engine to manage wheeled mode.
Most of the config is now to be found in the config.h file.

Deleted the old nuke.h, nuke.cpp and gait.cpp as the are now part of the gait_engine and inverse_kinematic files.

Started on a finite state machine for transitioning between modes but though better of it and commented the whole thing out to commit a functioning version before.